+++ contact-forms-anti-spam/includes/forms/elementor.php
@@ -101,7 +101,7 @@ function maspik_validation_process_elementor( $record, $ajax_handler ) {
   // General Check
   if(!$spam){
     $meta = $record->get_form_meta( [ 'page_url', 'remote_ip' ] );
-    $ip =  $meta['remote_ip']['value'] ? $meta['remote_ip']['value'] : efas_getRealIpAddr();
+    $ip =  $meta['remote_ip']['value'] ? $meta['remote_ip']['value'] : maspik_get_real_ip();
     // Country IP Check 
     $GeneralCheck = GeneralCheck($ip,$spam,$reason,$_POST,"elementor");
     $spam = isset($GeneralCheck['spam']) ? $GeneralCheck['spam'] : false;