+++ contact-forms-anti-spam/includes/forms/fluentforms.php
@@ -12,7 +12,7 @@ function maspik_validate_fluentform_general( $errors, $formData, $form, $fields)
   $spam = false;
   $reason ="";
   // ip
-  $ip =  efas_getRealIpAddr();
+  $ip =  maspik_get_real_ip();
 
 // For HP
 parse_str($_POST['data'], $parsed_data);
@@ -134,17 +134,6 @@ function maspik_validate_fluentforms_textarea($errorMessage, $field, $formData, 
 }
 add_filter('fluentform/validate_input_item_textarea', 'maspik_validate_fluentforms_textarea', 10, 5);
 
-
-// maspik_add_text_to_mail_components fluentforms
-//add_filter('fluentform/email_template_footer_text', 'maspik_add_text_to_mail_fluentforms', 10, 3);
-function maspik_add_text_to_mail_fluentforms($footerText, $form, $notification) {
-  $add_country_to_emails = maspik_get_settings("add_country_to_emails", '', 'old')  == "yes";
-  if($footerText && $add_country_to_emails){
-     $countryName = maspik_add_country_to_submissions($linebreak = "");
-     $footerText = $footerText.$countryName;
-    }
- return $footerText;
-}
 
 
 // add Maspik Honeypot fields to fluentform